+++ src/main.js
@@ -1,5 +1,5 @@
 // Description:
-//   A simple hubot that uses the mtg-sdk to query for specific cards.
+//   A simple hubot that uses the mtg-sdk (and scryfall-sdk) to query for specific cards.
 //
 // Commands:
 //   hubot mtg query [search param1]=[search value1],[search value2]|[search value3] & [search param2]= ...
@@ -7,6 +7,7 @@
 //   hubot mtg random
 //   hubot mtg clash [@handler]
 //   hubot mtg term [term]
+//   [[card name]] to display an image of the card
 //
 // Author:
 //   HerrPfister
@@ -15,7 +16,7 @@ var clash = require('./commands/clash'),
     findCard = require('./commands/find'),
     queryCards = require('./commands/query'),
     randomCard = require('./commands/random'),
-    scry = require('./commands/scryfall_find'),
+    scry = require('./commands/scry'),
     term = require('./commands/term');
 
 module.exports = function (robot) {
@@ -24,5 +25,5 @@ module.exports = function (robot) {
     robot.respond(/mtg\s+find\s+(.*)/i, findCard);
     robot.respond(/mtg\s+term\s+(.*)/i, term);
     robot.respond(/mtg\s+random/i, randomCard);
-    robot.hear(/[[(.*)]]/i, scry);
+    robot.hear(/\[\[(.*)\]\]/i, scry);
 };
